Extract LayoutWithGuidedTour
